move tag serializers to provider specific files
