+++ koku/api/tags/all/openshift/view.py
@@ -5,7 +5,7 @@
 """View for OCP-on-All tags."""
 from api.common.permissions.openshift_all_access import OpenshiftAllAccessPermission
 from api.tags.all.openshift.queries import OCPAllTagQueryHandler
-from api.tags.serializers import OCPAllTagsQueryParamSerializer
+from api.tags.all.openshift.serializers import OCPAllTagsQueryParamSerializer
 from api.tags.view import TagView
 from reporting.provider.aws.models import AWSTagsSummary
 from reporting.provider.azure.models import AzureTagsSummary
