+++ koku/api/tags/aws/serializers.py
@@ -1,0 +1,50 @@
+#
+# Copyright 2022 Red Hat Inc.
+# SPDX-License-Identifier: Apache-2.0
+#
+from rest_framework import serializers
+
+from api.report.serializers import add_operator_specified_fields
+from api.report.serializers import StringOrListField
+from api.tags.serializers import ExcludeSerializer
+from api.tags.serializers import FilterSerializer
+from api.tags.serializers import TagsQueryParamSerializer
+
+
+AWS_FILTER_OP_FIELDS = ["account", "org_unit_id"]
+
+
+class AWSFilterSerializer(FilterSerializer):
+    """Serializer for handling tag query parameter filter."""
+
+    account = StringOrListField(child=serializers.CharField(), required=False)
+    org_unit_id = StringOrListField(child=serializers.CharField(), required=False)
+    enabled = serializers.BooleanField(default=True, required=False)
+
+    def __init__(self, *args, **kwargs):
+        """Initialize the AWSFilterSerializer."""
+        super().__init__(*args, **kwargs)
+        add_operator_specified_fields(self.fields, AWS_FILTER_OP_FIELDS)
+
+
+class AWSExcludeSerializer(ExcludeSerializer):
+    """Serializer for handling tag query parameter exclude."""
+
+    account = StringOrListField(child=serializers.CharField(), required=False)
+    org_unit_id = StringOrListField(child=serializers.CharField(), required=False)
+    enabled = serializers.BooleanField(default=True, required=False)
+
+    def __init__(self, *args, **kwargs):
+        """Initialize the AWSExcludeSerializer."""
+        super().__init__(*args, **kwargs)
+        add_operator_specified_fields(self.fields, AWS_FILTER_OP_FIELDS)
+
+
+class AWSTagsQueryParamSerializer(TagsQueryParamSerializer):
+    """Serializer for handling AWS tag query parameters."""
+
+    EXCLUDE_SERIALIZER = AWSExcludeSerializer
+    FILTER_SERIALIZER = AWSFilterSerializer
+
+    exclude = AWSExcludeSerializer(required=False)
+    filter = AWSFilterSerializer(required=False)
