+++ koku/api/tags/azure/openshift/view.py
@@ -6,7 +6,7 @@
 from api.common.permissions.azure_access import AzureAccessPermission
 from api.common.permissions.openshift_access import OpenShiftAccessPermission
 from api.tags.azure.openshift.queries import OCPAzureTagQueryHandler
-from api.tags.serializers import OCPAzureTagsQueryParamSerializer
+from api.tags.azure.openshift.serializers import OCPAzureTagsQueryParamSerializer
 from api.tags.view import TagView
 from reporting.provider.azure.models import AzureTagsSummary
