+++ koku/api/tags/azure/serializers.py
@@ -1,0 +1,48 @@
+#
+# Copyright 2022 Red Hat Inc.
+# SPDX-License-Identifier: Apache-2.0
+#
+from rest_framework import serializers
+
+from api.report.serializers import add_operator_specified_fields
+from api.report.serializers import StringOrListField
+from api.tags.serializers import ExcludeSerializer
+from api.tags.serializers import FilterSerializer
+from api.tags.serializers import TagsQueryParamSerializer
+
+
+AZURE_FILTER_OP_FIELDS = ["subscription_guid"]
+
+
+class AzureFilterSerializer(FilterSerializer):
+    """Serializer for handling tag query parameter filter."""
+
+    subscription_guid = StringOrListField(child=serializers.CharField(), required=False)
+    enabled = serializers.BooleanField(default=True, required=False)
+
+    def __init__(self, *args, **kwargs):
+        """Initialize the AzureFilterSerializer."""
+        super().__init__(*args, **kwargs)
+        add_operator_specified_fields(self.fields, AZURE_FILTER_OP_FIELDS)
+
+
+class AzureExcludeSerializer(ExcludeSerializer):
+    """Serializer for handling tag query parameter exclude."""
+
+    subscription_guid = StringOrListField(child=serializers.CharField(), required=False)
+    enabled = serializers.BooleanField(default=True, required=False)
+
+    def __init__(self, *args, **kwargs):
+        """Initialize the AzureExcludeSerializer."""
+        super().__init__(*args, **kwargs)
+        add_operator_specified_fields(self.fields, AZURE_FILTER_OP_FIELDS)
+
+
+class AzureTagsQueryParamSerializer(TagsQueryParamSerializer):
+    """Serializer for handling Azure tag query parameters."""
+
+    EXCLUDE_SERIALIZER = AzureExcludeSerializer
+    FILTER_SERIALIZER = AzureFilterSerializer
+
+    exclude = AzureExcludeSerializer(required=False)
+    filter = AzureFilterSerializer(required=False)
