+++ koku/api/tags/gcp/openshift/view.py
@@ -6,7 +6,7 @@
 from api.common.permissions.gcp_access import GcpAccessPermission
 from api.common.permissions.openshift_access import OpenShiftAccessPermission
 from api.tags.gcp.openshift.queries import OCPGCPTagQueryHandler
-from api.tags.serializers import OCPGCPTagsQueryParamSerializer
+from api.tags.gcp.openshift.serializers import OCPGCPTagsQueryParamSerializer
 from api.tags.view import TagView
 from reporting.provider.gcp.models import GCPTagsSummary
