+++ koku/api/tags/ocp/view.py
@@ -5,7 +5,7 @@
 """View for OpenShift tags."""
 from api.common.permissions.openshift_access import OpenShiftAccessPermission
 from api.tags.ocp.queries import OCPTagQueryHandler
-from api.tags.serializers import OCPTagsQueryParamSerializer
+from api.tags.ocp.serializers import OCPTagsQueryParamSerializer
 from api.tags.view import TagView
 from reporting.provider.ocp.models import OCPStorageVolumeLabelSummary
 from reporting.provider.ocp.models import OCPUsagePodLabelSummary
